build: add -Wall and fix warnings

+++ sw/src/main.cpp
@@ -19,7 +19,7 @@ int main(int argc, char* argv[]) {
     findWordsInLine(" a[0] a[1] a[2] a[3] a[4] a[5] a[6] a[7] a[8] a[9] a[10] a[11] ");
     Graph graph;
     parseEQN(file_path, graph);
-    for (int i = 0; i < graph.nodes.size(); i++) {
+    for (size_t i = 0; i < graph.nodes.size(); i++) {
         cout << i << " ";
         graph.nodes[i].print();
         cout << endl;
